+++ python/src/services/user_repo.py
@@ -11,20 +11,27 @@ class UserRepo(IUserRepo):
     users: list[PeeweeUser] = PeeweeUser.select()
     return [x.to_interface() for x in users]
 
-  async def pick(self, where: IUserRepo.Where) -> RepoUser:
-    user: PeeweeUser = PeeweeUser.get(PeeweeUser.id == where.id)
+  async def pick(self, where: IUserRepo.Where) -> RepoUser | None:
+    user: PeeweeUser | None = PeeweeUser.get(PeeweeUser.id == where.id)
+    if user is None:
+      return None
     return user.to_interface()
 
-  async def create(self, data: IUserRepo.CreateData) -> None:
-    PeeweeUser.create(
+  async def create(self, data: IUserRepo.CreateData) -> RepoUser:
+    user: PeeweeUser = PeeweeUser.create(
       name=data.name,
     )
+    return user.to_interface()
 
   async def update(self, where: IUserRepo.Where, data: IUserRepo.UpdateData) -> None:
-    user: PeeweeUser = PeeweeUser.get(PeeweeUser.id == where.id)
+    user: PeeweeUser | None = PeeweeUser.get(PeeweeUser.id == where.id)
+    if user is None:
+      raise Exception("User not found")
     user.name = data.name
     user.save()
 
   async def delete(self, where: IUserRepo.Where) -> None:
-    user: PeeweeUser = PeeweeUser.get(PeeweeUser.id == where.id)
+    user: PeeweeUser | None = PeeweeUser.get(PeeweeUser.id == where.id)
+    if user is None:
+      raise Exception("User not found")
     user.delete_instance()
